+++ tests/test_callback.py
@@ -39,7 +39,7 @@ class CustomCallback(Callback):
     def on_next_configurations_start(self, smbo: smac.main.BaseSMBO) -> None:
         self.next_configurations_start_counter += 1
 
-    def on_next_configurations_end(self, smbo: smac.main.BaseSMBO, configurations: list[Configuration]) -> None:
+    def on_next_configurations_end(self, smbo: smac.main.BaseSMBO) -> None:
         self.next_configurations_end_counter += 1
 
     def on_ask_start(self, smbo: smac.main.BaseSMBO) -> None: